+++ test/analyzer/index_test.cpp
@@ -575,27 +575,27 @@ TEST_F(IndexTest, globalImport) {
                                                  R"(
 var _AAA = 34
 var AAA = $_AAA
-function BBB() : Int { return $AAA; }
+function BBB(aa: Int?) : Int { return $AAA; }
 CCC() { $BBB(); }
 typedef DDD = typeof(CCC)
 function EEE() for Int {}  # invalid method definition
-typedef FFF() { typedef GGG = Error; }
-function HHH() for FFF {}
+typedef FFF(bb:String?) { typedef GGG = Error; }
+function HHH(cc: Int?) for FFF {}
 )");
 
   unsigned short modId;
   auto content = format(R"(
 source %s
-$AAA + $BBB()
+$AAA + $BBB($aa: 2345)
 CCC
 new [DDD]()
 23.EEE()  # invalid method
 new FFF.GGG('34')
-new FFF().HHH()
+new FFF($bb:'').HHH($cc:22)
 )",
                         fileName.c_str());
   ASSERT_NO_FATAL_FAILURE(
-      this->doAnalyze(content.c_str(), modId, {.declSize = 0, .symbolSize = 8}));
+      this->doAnalyze(content.c_str(), modId, {.declSize = 0, .symbolSize = 11}));
   ASSERT_EQ(1, modId);
 
   // definition
@@ -604,29 +604,39 @@ new FFF().HHH()
   ASSERT_NO_FATAL_FAILURE(this->findDecl(
       Request{.modId = modId, .position = {.line = 2, .character = 9}}, {{2, "(3:9~3:12)"}}));
   ASSERT_NO_FATAL_FAILURE(this->findDecl(
+      Request{.modId = modId, .position = {.line = 2, .character = 14}}, {{2, "(3:13~3:15)"}}));
+  ASSERT_NO_FATAL_FAILURE(this->findDecl(
       Request{.modId = modId, .position = {.line = 3, .character = 1}}, {{2, "(4:0~4:3)"}}));
   ASSERT_NO_FATAL_FAILURE(this->findDecl(
       Request{.modId = modId, .position = {.line = 4, .character = 7}}, {{2, "(5:8~5:11)"}}));
   ASSERT_NO_FATAL_FAILURE(this->findDecl(
       Request{.modId = modId, .position = {.line = 6, .character = 6}}, {{2, "(7:8~7:11)"}}));
   ASSERT_NO_FATAL_FAILURE(this->findDecl(
-      Request{.modId = modId, .position = {.line = 6, .character = 8}}, {{2, "(7:24~7:27)"}}));
+      Request{.modId = modId, .position = {.line = 6, .character = 8}}, {{2, "(7:34~7:37)"}}));
   ASSERT_NO_FATAL_FAILURE(this->findDecl(
       Request{.modId = modId, .position = {.line = 7, .character = 6}}, {{2, "(7:8~7:11)"}}));
   ASSERT_NO_FATAL_FAILURE(this->findDecl(
-      Request{.modId = modId, .position = {.line = 7, .character = 12}}, {{2, "(8:9~8:12)"}}));
+      Request{.modId = modId, .position = {.line = 7, .character = 9}}, {{2, "(7:12~7:14)"}}));
+  ASSERT_NO_FATAL_FAILURE(this->findDecl(
+      Request{.modId = modId, .position = {.line = 7, .character = 18}}, {{2, "(8:9~8:12)"}}));
+  ASSERT_NO_FATAL_FAILURE(this->findDecl(
+      Request{.modId = modId, .position = {.line = 7, .character = 22}}, {{2, "(8:13~8:15)"}}));
 
   // references
   ASSERT_NO_FATAL_FAILURE(
       this->findRefs(Request{.modId = 2, .position = {.line = 2, .character = 5}},
                      {{2, "(2:4~2:7)"},    // itself
-                      {2, "(3:30~3:34)"},  // return $AAA;
+                      {2, "(3:38~3:42)"},  // return $AAA;
                       {1, "(2:0~2:4)"}})); // $AAA + $BBB()
   ASSERT_NO_FATAL_FAILURE(
       this->findRefs(Request{.modId = 2, .position = {.line = 3, .character = 10}},
                      {{2, "(3:9~3:12)"},    // itself
                       {2, "(4:8~4:12)"},    // $BBB()
                       {1, "(2:7~2:11)"}})); // $AAA + $BBB()
+  ASSERT_NO_FATAL_FAILURE(
+      this->findRefs(Request{.modId = 2, .position = {.line = 3, .character = 13}},
+                     {{2, "(3:13~3:15)"},    // itself
+                      {1, "(2:12~2:15)"}})); // $AAA + $BBB($aa: 2345)
   ASSERT_NO_FATAL_FAILURE(
       this->findRefs(Request{.modId = 2, .position = {.line = 4, .character = 2}},
                      {{2, "(4:0~4:3)"},    // itself
@@ -639,13 +649,21 @@ new FFF().HHH()
   ASSERT_NO_FATAL_FAILURE(
       this->findRefs(Request{.modId = 2, .position = {.line = 7, .character = 10}},
                      {{2, "(7:8~7:11)"},   // itself
-                      {2, "(8:19~8:22)"},  // function HHH() for FFF {}
+                      {2, "(8:27~8:30)"},  // function HHH() for FFF {}
                       {1, "(6:4~6:7)"},    // new FFF.GGG('34')
                       {1, "(7:4~7:7)"}})); // new FFF.HH()
   ASSERT_NO_FATAL_FAILURE(
-      this->findRefs(Request{.modId = 2, .position = {.line = 7, .character = 24}},
-                     {{2, "(7:24~7:27)"},   // itself
+      this->findRefs(Request{.modId = 2, .position = {.line = 7, .character = 13}},
+                     {{2, "(7:12~7:14)"},   // itself
+                      {1, "(7:8~7:11)"}})); // new FFF($bb:'')
+  ASSERT_NO_FATAL_FAILURE(
+      this->findRefs(Request{.modId = 2, .position = {.line = 7, .character = 36}},
+                     {{2, "(7:34~7:37)"},   // itself
                       {1, "(6:8~6:11)"}})); // new FFF.GGG('34')
+  ASSERT_NO_FATAL_FAILURE(
+      this->findRefs(Request{.modId = 2, .position = {.line = 8, .character = 14}},
+                     {{2, "(8:13~8:15)"},    // itself
+                      {1, "(7:20~7:23)"}})); // new FFF($bb:'').HHH($cc:22)
 }
 
 TEST_F(IndexTest, namedImport) {
@@ -654,27 +672,27 @@ TEST_F(IndexTest, namedImport) {
                                                  R"(
 var _AAA = 34
 var AAA = $_AAA
-function BBB() : Int { return $AAA; }
+function BBB(aa:Int) : Int { return $AAA; }
 CCC() { $BBB(); }
 typedef DDD = typeof(CCC)
 function EEE() : Int for Int { return 34; }  #  invalid method definition
-typedef FFF() { var value = 34; }
-function HHH() for FFF {}
+typedef FFF(bb:Int?) { var value = 34; }
+function HHH(cc:Int?) for FFF {}
 )");
 
   unsigned short modId;
   auto content = format(R"(source %s \
 as mod
-$mod.AAA + $mod.BBB()
+$mod.AAA + $mod.BBB($aa:44)
 mod 2>&1 > /dev/null CCC 34
 new [mod.DDD]()
 23.EEE()  # invalid method call
-new mod.FFF().value
-new mod.FFF().HHH()
+new mod.FFF($bb:32).value
+new mod.FFF($bb:11).HHH($cc:2)
 )",
                         fileName.c_str());
   ASSERT_NO_FATAL_FAILURE(
-      this->doAnalyze(content.c_str(), modId, {.declSize = 1, .symbolSize = 15}));
+      this->doAnalyze(content.c_str(), modId, {.declSize = 1, .symbolSize = 19}));
   ASSERT_EQ(1, modId);
 
   // definition
@@ -687,17 +705,23 @@ new mod.FFF().HHH()
   ASSERT_NO_FATAL_FAILURE(this->findDecl(
       Request{.modId = modId, .position = {.line = 2, .character = 18}}, {{2, "(3:9~3:12)"}}));
   ASSERT_NO_FATAL_FAILURE(this->findDecl(
+      Request{.modId = modId, .position = {.line = 2, .character = 22}}, {{2, "(3:13~3:15)"}}));
+  ASSERT_NO_FATAL_FAILURE(this->findDecl(
       Request{.modId = modId, .position = {.line = 3, .character = 23}}, {{2, "(4:0~4:3)"}}));
   ASSERT_NO_FATAL_FAILURE(this->findDecl(
       Request{.modId = modId, .position = {.line = 4, .character = 10}}, {{2, "(5:8~5:11)"}}));
   ASSERT_NO_FATAL_FAILURE(this->findDecl(
       Request{.modId = modId, .position = {.line = 6, .character = 10}}, {{2, "(7:8~7:11)"}}));
   ASSERT_NO_FATAL_FAILURE(this->findDecl(
-      Request{.modId = modId, .position = {.line = 6, .character = 18}}, {{2, "(7:20~7:25)"}}));
+      Request{.modId = modId, .position = {.line = 6, .character = 13}}, {{2, "(7:12~7:14)"}}));
+  ASSERT_NO_FATAL_FAILURE(this->findDecl(
+      Request{.modId = modId, .position = {.line = 6, .character = 22}}, {{2, "(7:27~7:32)"}}));
   ASSERT_NO_FATAL_FAILURE(this->findDecl(
       Request{.modId = modId, .position = {.line = 7, .character = 10}}, {{2, "(7:8~7:11)"}}));
   ASSERT_NO_FATAL_FAILURE(this->findDecl(
-      Request{.modId = modId, .position = {.line = 7, .character = 16}}, {{2, "(8:9~8:12)"}}));
+      Request{.modId = modId, .position = {.line = 7, .character = 22}}, {{2, "(8:9~8:12)"}}));
+  ASSERT_NO_FATAL_FAILURE(this->findDecl(
+      Request{.modId = modId, .position = {.line = 7, .character = 26}}, {{2, "(8:13~8:15)"}}));
 
   // references
   ASSERT_NO_FATAL_FAILURE(
@@ -711,12 +735,16 @@ new mod.FFF().HHH()
                       {modId, "(7:4~7:7)"}}));
   ASSERT_NO_FATAL_FAILURE(this->findRefs(
       Request{.modId = 2, .position = {.line = 2, .character = 5}}, {{2, "(2:4~2:7)"}, // itself
-                                                                     {2, "(3:30~3:34)"},
+                                                                     {2, "(3:36~3:40)"},
                                                                      {1, "(2:5~2:8)"}}));
   ASSERT_NO_FATAL_FAILURE(this->findRefs(
       Request{.modId = 2, .position = {.line = 3, .character = 10}}, {{2, "(3:9~3:12)"}, // itself
                                                                       {2, "(4:8~4:12)"},
                                                                       {1, "(2:16~2:19)"}}));
+  ASSERT_NO_FATAL_FAILURE(
+      this->findRefs(Request{.modId = 2, .position = {.line = 3, .character = 14}},
+                     {{2, "(3:13~3:15)"},    // itself
+                      {1, "(2:20~2:23)"}})); // $mod.BBB($aa:44)
   ASSERT_NO_FATAL_FAILURE(this->findRefs(
       Request{.modId = 2, .position = {.line = 4, .character = 2}}, {{2, "(4:0~4:3)"}, // itself
                                                                      {2, "(5:21~5:24)"},
@@ -727,17 +755,26 @@ new mod.FFF().HHH()
   ASSERT_NO_FATAL_FAILURE(
       this->findRefs(Request{.modId = 2, .position = {.line = 7, .character = 10}},
                      {{2, "(7:8~7:11)"},    // itself
-                      {2, "(8:19~8:22)"},   // function HHH() for FFF {}
+                      {2, "(8:26~8:29)"},   // function HHH() for FFF {}
                       {1, "(6:8~6:11)"},    // new mod.FFF().value
                       {1, "(7:8~7:11)"}})); // new mod.FFF().HHH()
+  ASSERT_NO_FATAL_FAILURE(
+      this->findRefs(Request{.modId = 2, .position = {.line = 7, .character = 12}},
+                     {{2, "(7:12~7:14)"},    // itself
+                      {1, "(6:12~6:15)"},    // new mod.FFF($bb:32).value
+                      {1, "(7:12~7:15)"}})); // new mod.FFF($bb:11).HHH()
   ASSERT_NO_FATAL_FAILURE(this->findRefs(
-      Request{.modId = 2, .position = {.line = 7, .character = 24}}, {{2, "(7:20~7:25)"}, // itself
-                                                                      {1, "(6:14~6:19)"}}));
+      Request{.modId = 2, .position = {.line = 7, .character = 30}}, {{2, "(7:27~7:32)"}, // itself
+                                                                      {1, "(6:20~6:25)"}}));
 
   ASSERT_NO_FATAL_FAILURE(
       this->findRefs(Request{.modId = 2, .position = {.line = 8, .character = 11}},
                      {{2, "(8:9~8:12)"},     // itself
-                      {1, "(7:14~7:17)"}})); // new mod.FFF().HHH()
+                      {1, "(7:20~7:23)"}})); // new mod.FFF().HHH()
+  ASSERT_NO_FATAL_FAILURE(
+      this->findRefs(Request{.modId = 2, .position = {.line = 8, .character = 14}},
+                     {{2, "(8:13~8:15)"},    // itself
+                      {1, "(7:24~7:27)"}})); // new mod.FFF($bb:11).HHH($cc:2)
 }
 
 TEST_F(IndexTest, namedImportInlined) {